requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.1
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
Flask-SQLAlchemy==2.4.0
future==0.17.1
geatpy==2.2.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.3.3
macholib==1.11
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.3
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pefile==2019.4.18
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.1
PyHamcrest==1.9.0
pyinstall==0.1.4
PyInstaller==3.4
pylint==2.4.4
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypiwin32==223
PyQt5==5.11.3
PyQt5-sip==4.19.17
pyqt5-tools==5.11.3.1.4
pyrsistent==0.15.2
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.5
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
sip==4.19.8
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
SQLAlchemy==1.3.3
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.4.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
zope.interface==4.6.0
